+++ changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- In `SingleInstanceTaskScheduler` constructor options, rename `nextRunTimeEvaluator`
+  to `nextRunTime` and accept the same options as `buildEvaluator`.
 
 ## [0.1.0] - 2021-05-24
 ### Added